fullcalendar updates

+++ web/frontend/calendar/delete-event-frontend.php
@@ -3,9 +3,12 @@ require_once './../core/calendar-init.php';
 
 $id = Input::post('id');
 
-$deleteEvent = $frontDb->delete(Params::TBL_EVENTS, ActionCond::where(['id', $id]));
-
-if ($deleteEvent) {
-    echo 1;
+if ($id !== 'undefined') {
+    $deleteEvent = $frontDb->delete(Params::TBL_EVENTS, ActionCond::where(['id', $id]));
+    if ($deleteEvent) {
+        echo 1;
+    }
+} else {
+    echo 0;
 }
 ?>
